move RebuildIndex to cairo
