+++ utils/src/main/java/io/questdb/cliutil/RebuildIndex.java
@@ -24,193 +24,3 @@
 
 package io.questdb.cliutil;
 
-import io.questdb.cairo.*;
-import io.questdb.cairo.vm.Vm;
-import io.questdb.cairo.vm.api.MemoryMAR;
-import io.questdb.cairo.vm.api.MemoryMR;
-import io.questdb.log.Log;
-import io.questdb.log.LogFactory;
-import io.questdb.std.FilesFacadeImpl;
-import io.questdb.std.MemoryTag;
-import io.questdb.std.datetime.DateFormat;
-import io.questdb.std.str.Path;
-import io.questdb.std.str.StringSink;
-
-import java.io.Closeable;
-
-import static io.questdb.cairo.TableUtils.lockName;
-
-public class RebuildIndex implements Closeable {
-    private final Path path;
-    private final int rootLen;
-    private final FilesFacadeImpl ff;
-    private final MemoryMAR ddlMem;
-    private final CairoConfiguration configuration;
-    private long lockFd;
-    private final MemoryMR indexMem = Vm.getMRInstance();
-    private static final Log LOG = LogFactory.getLog(RebuildIndex.class);
-
-    public RebuildIndex(String tablePath, CairoConfiguration configuration) {
-        this.path = new Path();
-        this.path.concat(tablePath);
-        this.rootLen = tablePath.length();
-        this.ff = new FilesFacadeImpl();
-        this.ddlMem = Vm.getMARInstance();
-        this.configuration = configuration;
-    }
-
-    public void rebuildAllPartitions() {
-        lock();
-        path.trimTo(rootLen);
-
-        SymbolColumnIndexer indexer = new SymbolColumnIndexer();
-
-        try (TableReaderMetadata metadata = new TableReaderMetadata(ff)) {
-            path.concat(TableUtils.META_FILE_NAME);
-            metadata.of(path, ColumnType.VERSION);
-            path.trimTo(rootLen);
-
-            int partitionBy = metadata.getPartitionBy();
-            DateFormat partitionDirFormatMethod = PartitionBy.getPartitionDirFormatMethod(partitionBy);
-            StringSink sink = new StringSink();
-
-            try (TxReader txReader = new TxReader(ff, path, partitionBy)) {
-                txReader.unsafeLoadAll();
-                path.trimTo(rootLen);
-
-                for (int columnIndex = metadata.getColumnCount() - 1; columnIndex > -1; columnIndex--) {
-                    if (metadata.isColumnIndexed(columnIndex)) {
-                        CharSequence columnName = metadata.getColumnName(columnIndex);
-
-                        if (PartitionBy.isPartitioned(partitionBy)) {
-                            for (int partitionIndex = txReader.getPartitionCount() - 1; partitionIndex > -1; partitionIndex--) {
-                                long partitionTimestamp = txReader.getPartitionTimestamp(partitionIndex);
-                                long partitionSize = txReader.getPartitionSize(partitionIndex);
-
-                                sink.clear();
-                                partitionDirFormatMethod.format(partitionTimestamp, null, null, sink);
-                                rebuildIndex(indexer, columnName, sink.toString(), metadata.getIndexValueBlockCapacity(columnIndex), partitionSize);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    private void rebuildIndex(SymbolColumnIndexer indexer, CharSequence columnName, String partitionName, int indexValueBlockCapacity, long partitionSize) {
-        path.trimTo(rootLen).concat(partitionName);
-        final int plen = path.length();
-
-        if (ff.exists(path)) {
-
-            try (final MemoryMR roMem = indexMem) {
-                LOG.info().$("indexing [path=").$(path).I$();
-
-                removeIndexFiles(columnName);
-
-                TableUtils.dFile(path.trimTo(plen), columnName);
-                createIndexFiles(columnName, indexValueBlockCapacity, plen, true);
-                final long columnTop = TableUtils.readColumnTop(ff, path.trimTo(plen), columnName, plen, false);
-
-                if (partitionSize > columnTop) {
-                    TableUtils.dFile(path.trimTo(plen), columnName);
-                    final long columnSize = (partitionSize - columnTop) << ColumnType.pow2SizeOf(ColumnType.INT);
-                    roMem.of(ff, path, columnSize, columnSize, MemoryTag.MMAP_TABLE_WRITER);
-                    indexer.configureWriter(configuration, path.trimTo(plen), columnName, columnTop);
-                    indexer.index(roMem, columnTop, partitionSize);
-                }
-            }
-        }
-
-    }
-
-    private void createIndexFiles(CharSequence columnName, int indexValueBlockCapacity, int plen, boolean force) {
-        try {
-            BitmapIndexUtils.keyFileName(path.trimTo(plen), columnName);
-
-            if (!force && ff.exists(path)) {
-                return;
-            }
-
-            // reuse memory column object to create index and close it at the end
-            try {
-                ddlMem.smallFile(ff, path, MemoryTag.MMAP_TABLE_WRITER);
-                BitmapIndexWriter.initKeyMemory(ddlMem, indexValueBlockCapacity);
-            } catch (CairoException e) {
-                // looks like we could not create key file properly
-                // lets not leave half-baked file sitting around
-                LOG.error()
-                        .$("could not create index [name=").utf8(path)
-                        .$(", errno=").$(e.getErrno())
-                        .$(']').$();
-                if (!ff.remove(path)) {
-                    LOG.error()
-                            .$("could not remove '").utf8(path).$("'. Please remove MANUALLY.")
-                            .$("[errno=").$(ff.errno())
-                            .$(']').$();
-                }
-                throw e;
-            } finally {
-                ddlMem.close();
-            }
-            if (!ff.touch(BitmapIndexUtils.valueFileName(path.trimTo(plen), columnName))) {
-                LOG.error().$("could not create index [name=").$(path).$(']').$();
-                throw CairoException.instance(ff.errno()).put("could not create index [name=").put(path).put(']');
-            }
-        } finally {
-            path.trimTo(plen);
-        }
-    }
-
-    private void removeIndexFiles(CharSequence columnName) {
-        final int plen = path.length();
-        BitmapIndexUtils.keyFileName(path.trimTo(plen), columnName);
-        removeFile(path);
-
-        BitmapIndexUtils.valueFileName(path.trimTo(plen), columnName);
-        removeFile(path);
-    }
-
-    private void removeFile(Path path) {
-        LOG.info().$("deleting ").$(path).$();
-        ff.remove(this.path);
-    }
-
-    private void lock() {
-        try {
-            path.trimTo(rootLen);
-            lockName(path);
-            this.lockFd = TableUtils.lock(ff, path);
-        } finally {
-            path.trimTo(rootLen);
-        }
-
-        if (this.lockFd == -1L) {
-            throw CairoException.instance(ff.errno()).put("Cannot lock table: ").put(path.$());
-        }
-    }
-
-    @Override
-    public void close() {
-        path.trimTo(rootLen);
-        lockName(path);
-        releaseLock();
-        this.path.close();
-    }
-
-    private void releaseLock() {
-        if (lockFd != -1L) {
-            ff.close(lockFd);
-            try {
-                path.trimTo(rootLen);
-                lockName(path);
-                if (ff.exists(path) && !ff.remove(path)) {
-                    throw CairoException.instance(ff.errno()).put("Cannot remove ").put(path);
-                }
-            } finally {
-                path.trimTo(rootLen);
-            }
-        }
-    }
-}
